Split model and other metric updates
